get_security_group/requirements.txt
requests
click
PyYAML
esdk-obs-python
huaweicloudsdkcore
huaweicloudsdknat
huaweicloudsdkiam
huaweicloudsdkvpc